Make it an integer

+++ R/s3.put.R
@@ -28,7 +28,7 @@ s3.put <- function (x, path, name, bucket.location = "US", verbose = FALSE,
           "--debug", ""), '--check-md5')
 
   ## Ensure backoff vector has correct number of elements and is capped
-  backoff <- backoff[vapply(1:num_retries, function(i) min(i, length(backoff)), numeric(1))]
+  backoff <- backoff[vapply(1:num_retries, function(i) min(i, length(backoff)), integer(1))]
   run_system_put(path, name, s3.cmd, check_exists, num_retries, backoff)
 }
